data_service/requirements.txt
Flask==2.3.2
python-dotenv
requests==2.26.0
PyJWT